[refactor]:Socket Chat도메인 관련 api 분리
