+++ socket/src/main/java/Backend/socket/api/room/controller/RoomController.java
@@ -1,20 +1,14 @@
-package Backend.socket.domain.chat.application.controller;
+package Backend.socket.api.room.controller;
 
-import Backend.socket.domain.chat.application.controller.dto.request.ChatListRequestDto;
-import Backend.socket.domain.chat.application.controller.dto.request.ChatMessageListRequestDto;
-import Backend.socket.domain.chat.application.controller.dto.request.ChatMessageRoomRequestDto;
+import Backend.socket.api.chat.service.dto.request.ChatMessageRoomRequestDto;
 import Backend.socket.domain.chat.application.controller.dto.response.*;
-import Backend.socket.domain.chat.application.service.ChatService;
+import Backend.socket.api.chat.service.ChatService;
 import Backend.socket.domain.chat.application.service.RoomService;
-import Backend.socket.domain.chat.domain.Chat;
 import Backend.socket.global.common.MessageSuccessCode;
 import Backend.socket.global.common.MessageSuccessResponse;
-import Backend.socket.infra.config.auth.UserId;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.data.redis.core.RedisTemplate;
-import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.handler.annotation.DestinationVariable;
-import org.springframework.messaging.handler.annotation.Header;
 import org.springframework.messaging.handler.annotation.MessageMapping;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.web.bind.annotation.PathVariable;
