--- socket/src/main/java/Backend/socket/domain/chat/application/controller/dto/request/ChatListRequestDto.java
@@ -1,11 +1,0 @@
-package Backend.socket.domain.chat.application.controller.dto.request;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class ChatListRequestDto {
-    private String userName;
-}
