--- socket/src/main/java/Backend/socket/domain/chat/application/controller/dto/request/ChatMessageListRequestDto.java
@@ -1,13 +1,0 @@
-package Backend.socket.domain.chat.application.controller.dto.request;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class ChatMessageListRequestDto {
-    private String chatSession;
-    private String fromUserName;
-    private String toUserName;
-}
