--- socket/src/main/java/Backend/socket/domain/chat/application/controller/dto/request/RoomMessageListRequestDto.java
@@ -1,13 +1,0 @@
-package Backend.socket.domain.chat.application.controller.dto.request;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-public class RoomMessageListRequestDto {
-    private String chatSession;
-    private String fromUserName;
-    private String toRoomName;
-}
